LayoutTests/imported/w3c/web-platform-tests/trusted-types/HTMLScriptElement-internal-slot-expected.txt

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.

CONSOLE MESSAGE: ReferenceError: Can't find variable: trustedTypes
CONSOLE MESSAGE: Unrecognized Content-Security-Policy directive 'require-trusted-types-for'.



Harness Error (FAIL), message = ReferenceError: Can't find variable: trustedTypes

FAIL Test TT application when manipulating <script> elements during loading. promise_test: Unhandled rejection with value: "First message should have been blocked: first script element executed"
